+++ resources/js/bootstrap.js
@@ -26,3 +26,5 @@ $('.select2').select2();
 
 import Quill from 'quill';
 window.Quill = Quill;
+
+import 'livewire-sortable'