+++ resources/views/components/dialog-modal.blade.php
@@ -1,6 +1,6 @@
 @props(['title', 'content', 'id' => null, 'maxWidth' => null, 'button' => null, 'show' => true, 'withoutFooter' => false])
 
-<x-move-modal :id="$id" :maxWidth="$maxWidth" {{ $attributes }} :show="$show"  :button="$button">
+<x-move-modal :id="$id" :maxWidth="$maxWidth" {{ $attributes }} :show="$show" :button="$button">
     <div class="px-6 py-4 text-left">
         <div class="text-lg">
             {{ $title }}