+++ resources/views/components/panel.blade.php
@@ -1,4 +1,10 @@
-@props(['title' => null, 'description' => null, 'panel' => null])
+@props(['title' => null, 'description' => null, 'panel' => null, 'overflowHidden'])
+
+@php
+
+$overflowHidden ??= $panel->overflowHidden;
+
+@endphp
 
 @if ($title || $description || $panel->description)
     @if ($title) <h2 class="{{ ($panel->class ?? null) ?: 'text-2xl font-semibold text-gray-900 mt-5' }}">{{ $title }}</h2> @endif
@@ -7,7 +13,7 @@
     @endif
 @endif
 
-<div {{ $attributes->merge(['class' => 'shadow overflow-hidden sm:rounded-md']) }}>
+<div {{ $attributes->merge(['class' => 'shadow '. ($overflowHidden ? 'overflow-hidden' : null) .' sm:rounded-md']) }}>
 
     <div class="bg-white">